requirements.txt
requests
beautifulsoup4
pyfiglet
nmap
whois
dnsresolver
termcolor

